Set babel compilation on the server
In anticipation to loading es2015 client side code soon.
